change add_vertex to accept Vertex
add_vertex used to accept a set of co-ordinates and a name, now it it accepts a vertex and a name. In effect, this allows the user to give the same vertex multiple names, making it easier to reference the same vertex in multiple contexts.

In addition, a primitive check for name collisions in the  `vertices` array has been added.