Comment invalid edges
Sometimes edges are created between 1 or 2 vertices which are not used in any blocks. When this happens, they are not assigned an Index, and blockMesh throws an error when it encounters 'None' where the index should be in the blockMeshDict. In these cases the edge is not useful anyway. To avoid these issues, the edge is written as a comment in blockMesh dict when either of its vertices have not been assigned an index.