Accomodate blockMesh bug
For some reason, blockMesh fails when an empty geometry section is written as:

geometry
(
);

while it accepts:

geometry();

However openfoam throws an error when it encounters:

boundary();

but not when I write:

boundary
(
);

This needs an extra `elif` clause to handle this case. I suspect the bug has something to do with the Windows / Linux line-endings differences.