Restructure Vertex Objects and usage
Point and Vertex objects are now initiated from array-like objects containing three co-ordinates, and a transformation function.

In the builder classes, the array passed to each of the 'baked_vertices' is a view of the structured array of vertices.

The transformation function is called just prior to writing each Point, to transfrom to the Cartesian co-ordinate system.

As a result, the `bake_vertices` functions have been scrapped.

All examples are updated accordingly.