Face and BlockMeshDict class updates
Change Face class so it can format both faces which are used in the boundary section, and faces which are to be projected onto geometries.

All the `add` methods no longer require a name to be passed.

Because of this, the faces list in the BlockMeshDict can share the general formatting function, `_format_section`.

BlockMeshDict now stores objects in arrays instead of dictionaries. There's no reason for mapping names to object lists when there's no reason to refer to them after writing.

As a consequence, naming is now optional for many classes. In the future, may omit these class members alltogether.