Restructured Face class
Face class initializer now requires a 2x2 array. This seems physically more intuitive, even if the code gets a little uglier. This change allows the creation of faces which depend on views of the structured array in the builder classes.

'faces' has been added as a field to the builder classes, and are initialized as expected.

An additional argument is passed to the face format section, to distinguish whether the face is being written as part of a boundary section, or whether it is to be projected onto a geometry.

`write_blocks` has been renamed to `write` to indicate that more than just blocks will be written.

Examples have been updated accordingly.

BUGFIX: fixed order of arguments in `ArcEdge` and `SplineEdge` Initializers.