Added eighth_twist option to CylBlockStructContainer
This makes it so the o-grid can be oriented either flush with the cardinal planes, or on a 45 degree angle. This is helpful for creating wedges of axisymmetric domains.

The option has been added to the cyl_struct example along with a tip for choosing the number of points in ts