Split code into sensible file structure
Previously, it was enough to keep code in 2 major files. These files have become more cumbersome, and it's time to break the code into more manageable pieces.

I debated whether to restrict the user code to access the packages in the same way as the new file structure. For example

`blockmeshbuilder.blockmeshdict.BlockMeshDict`

After trying out this technique, I decided it didn't make sense, because the user code is likely to use classes from multiple different submodules most of the time.

For the moment, all necessary objects have been pulled into the 'root' level of the module so that user code is unaffected by this commit.